+++ src/main/java/insane96mcp/runeenchanting/runes/FeatherFallingRune.java
@@ -1,12 +1,20 @@
 package insane96mcp.runeenchanting.runes;
 
 import insane96mcp.insanelib.core.feature.config.Config;
+import net.minecraft.data.tags.IntrinsicHolderTagsProvider;
 import net.minecraft.tags.DamageTypeTags;
+import net.minecraft.tags.ItemTags;
 import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.item.Item;
 
 public class FeatherFallingRune extends DamageReductionRune {
     @Config
     public static Double damageReduction = 0.75d;
+
+    @Override
+    public void addItemsToApplicableTag(IntrinsicHolderTagsProvider.IntrinsicTagAppender<Item> appender) {
+        appender.addTag(ItemTags.FOOT_ARMOR_ENCHANTABLE);
+    }
 
     @Override
     public String getName() {