+++ changelog.md
@@ -1,3 +1,6 @@
+# Upcoming
+* Fixed Feather Falling applicable to any armor piece
+
 # 1.0.1.3-beta
 * Rune items now show which items they're compatible with in their tooltip, based on the `runeenchanting:display_on_rune` item tag