+++ changelog.md
@@ -3,6 +3,7 @@
   * Now always report best values and not current or stacked ones
     * E.g. Vindication no longer shows accumulated damage
 * Veining and Tunneling Runes' mining speed penalty is now configurable
+* Slightly increased Thorns chance to trigger
 * Fixed Feather Falling applicable to any armor piece
 
 # 1.0.1.3-beta